% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{messari_reprex}
\alias{messari_reprex}
\title{Prices of 50,000 round cut diamonds.}
\format{
A data frame with 1369627 rows and 83 variables:
\describe{
  \item{DateExtracted}{Date the data was collected on - Colorado Timezone}
  \item{DateTimeColoradoMDT}{Date and Time the data was collected on - Colorado Timezone}
  \item{Name}{The name of the cryptocurrency}
  \item{Rank}{The rank in terms of market capitalization of the cryptocurrency at the time the data was collected}
  \item{PriceUSD}{Price in US Dollars}
  ...
}
}
\source{
\url{https://messari.io/}
}
\usage{
messari_reprex
}
\description{
A dataset containing cryptocurrency data from messari.io
}
\keyword{datasets}
